Create Swag Calculator
This is a swag calculator. Its arguments are your name, the car you drive and monetary value. I have displayed good syntax and the way in which to write a testing class here. For anyone who uses this in the future please be cognizant enough to change the context and/or variable/method names.