Create Sudoku.java
We had to add this in order to be able all the data types in the project guidelines IE the floats, doubles and bytes. In the end we just ended up immediately casting all these stupid variables that we created right back into ints and into the then original ints in this sudoku program.